api/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.4.4
plum-py==0.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2019.3
SQLAlchemy==1.3.23
virtualenv==20.4.2
Werkzeug==1.0.1
geoalchemy2==0.8.5
pillow==7.2.0
